+++ src/main/java/seedu/recipe/MainApp.java
@@ -26,14 +26,14 @@ import seedu.recipe.model.UserPrefs;
 import seedu.recipe.model.cooked.CookedRecordBook;
 import seedu.recipe.model.plan.PlannedBook;
 import seedu.recipe.model.util.SampleDataUtil;
-import seedu.recipe.storage.CookedRecordBookStorage;
-import seedu.recipe.storage.JsonCookedRecordBookStorage;
 import seedu.recipe.storage.JsonRecipeBookStorage;
 import seedu.recipe.storage.JsonUserPrefsStorage;
 import seedu.recipe.storage.RecipeBookStorage;
 import seedu.recipe.storage.Storage;
 import seedu.recipe.storage.StorageManager;
 import seedu.recipe.storage.UserPrefsStorage;
+import seedu.recipe.storage.cooked.CookedRecordBookStorage;
+import seedu.recipe.storage.cooked.JsonCookedRecordBookStorage;
 import seedu.recipe.storage.plan.JsonPlannedBookStorage;
 import seedu.recipe.storage.plan.PlannedBookStorage;
 import seedu.recipe.ui.Ui;
