+++ src/test/java/seedu/recipe/storage/StorageManagerTest.java
@@ -14,6 +14,7 @@ import seedu.recipe.commons.core.GuiSettings;
 import seedu.recipe.model.ReadOnlyRecipeBook;
 import seedu.recipe.model.RecipeBook;
 import seedu.recipe.model.UserPrefs;
+import seedu.recipe.storage.cooked.JsonCookedRecordBookStorage;
 import seedu.recipe.storage.plan.JsonPlannedBookStorage;
 
 public class StorageManagerTest {
